+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.8.0
 - Autoclear to optionally reset state after period of time
+- Validate regular expressions at startup for correcting plates
 # 0.7.2
 - Python moved to 3.14 for docker image
 ## Home Assistant Integration